Visualize RM Vocabulary using Jupypter notebook
Visualize the RM Vocabulary, which wonderfully plots the concepts in
the vocabulary, showing their relatedness in the RM hierarchy. This
visualization was demonstrated at the ENVRI Week at Zandvoort on 14
May. This initial commit was a Jupyter notebook of the code and the
output, which can be shared via Github and NbViewer.